refactor: split agent tools and dispatch into per-service files
Split the monolithic tools.go and agent.go into per-service files to
reduce merge conflicts when multiple agents work on different services
in parallel. Tool definitions and input structs move to tools_*.go,
dispatch logic moves to dispatch_*.go. No behavioral changes.
