+++ internal/agent/agent.go
@@ -286,518 +286,22 @@ func sanitizeInput(raw json.RawMessage) string {
 
 // dispatchTool executes a tool call and returns the JSON result string and whether it's an error.
 func (a *Agent) dispatchTool(ctx context.Context, name string, rawInput json.RawMessage) (string, bool) {
-	var result any
-	var err error
-
-	switch name {
-	case "search_series":
-		var input searchSeriesInput
-		if err := json.Unmarshal(rawInput, &input); err != nil {
-			return jsonError("invalid input: " + err.Error()), true
-		}
-		result, err = a.sonarr.Search(ctx, input.Term)
-
-	case "add_series":
-		var input addSeriesInput
-		if err := json.Unmarshal(rawInput, &input); err != nil {
-			return jsonError("invalid input: " + err.Error()), true
-		}
-		result, err = a.sonarr.Add(ctx, sonarr.AddSeriesRequest{
-			Title:            input.Title,
-			TVDBID:           input.TVDBID,
-			QualityProfileID: input.QualityProfileID,
-			RootFolderPath:   input.RootFolderPath,
-			Monitored:        true,
-			SeasonFolder:     true,
-		})
-
-	case "get_queue":
-		result, err = a.sonarr.Queue(ctx)
-
-	case "get_history":
-		var input getHistoryInput
-		if err := json.Unmarshal(rawInput, &input); err != nil {
-			return jsonError("invalid input: " + err.Error()), true
-		}
-		pageSize := input.PageSize
-		if pageSize == 0 {
-			pageSize = 20
-		}
-		result, err = a.sonarr.History(ctx, pageSize)
-
-	case "check_health":
-		result, err = a.sonarr.Health(ctx)
-
-	case "remove_failed":
-		var input removeFailedInput
-		if err := json.Unmarshal(rawInput, &input); err != nil {
-			return jsonError("invalid input: " + err.Error()), true
-		}
-		err = a.sonarr.RemoveFailed(ctx, input.ID, input.Blocklist)
-		if err == nil {
-			result = map[string]string{"status": "removed"}
-		}
-
-	case "get_series_detail":
-		var input getSeriesDetailInput
-		if err := json.Unmarshal(rawInput, &input); err != nil {
-			return jsonError("invalid input: " + err.Error()), true
-		}
-		result, err = a.sonarr.GetSeries(ctx, input.SeriesID)
-
-	case "get_episodes":
-		var input getEpisodesInput
-		if err := json.Unmarshal(rawInput, &input); err != nil {
-			return jsonError("invalid input: " + err.Error()), true
-		}
-		result, err = a.sonarr.GetEpisodes(ctx, input.SeriesID)
-
-	case "get_logs":
-		var input getLogsInput
-		if err := json.Unmarshal(rawInput, &input); err != nil {
-			return jsonError("invalid input: " + err.Error()), true
-		}
-		result, err = a.sonarr.GetLogs(ctx, input.PageSize, input.Level)
-
-	case "manual_search":
-		var input manualSearchInput
-		if err := json.Unmarshal(rawInput, &input); err != nil {
-			return jsonError("invalid input: " + err.Error()), true
-		}
-		result, err = a.sonarr.ManualSearch(ctx, input.EpisodeID)
-
-	case "get_quality_profiles":
-		result, err = a.sonarr.GetQualityProfiles(ctx)
-
-	case "get_blocklist":
-		var input getBlocklistInput
-		if err := json.Unmarshal(rawInput, &input); err != nil {
-			return jsonError("invalid input: " + err.Error()), true
-		}
-		result, err = a.sonarr.GetBlocklist(ctx, input.PageSize)
-
-	case "get_root_folders":
-		result, err = a.sonarr.GetRootFolders(ctx)
-
-	case "get_download_clients":
-		result, err = a.sonarr.GetDownloadClients(ctx)
-
-	case "update_series_monitoring":
-		var input updateSeriesMonitoringInput
-		if err := json.Unmarshal(rawInput, &input); err != nil {
-			return jsonError("invalid input: " + err.Error()), true
-		}
-		series, getErr := a.sonarr.GetSeries(ctx, input.SeriesID)
-		if getErr != nil {
-			return jsonError(getErr.Error()), true
-		}
-		found := false
-		for i, s := range series.Seasons {
-			if s.SeasonNumber == input.SeasonNumber {
-				series.Seasons[i].Monitored = input.Monitored
-				found = true
-				break
-			}
-		}
-		if !found {
-			return jsonError(fmt.Sprintf("season %d not found in series %q", input.SeasonNumber, series.Title)), true
-		}
-		result, err = a.sonarr.UpdateSeries(ctx, series)
-
-	case "trigger_series_search":
-		var input triggerSeriesSearchInput
-		if err := json.Unmarshal(rawInput, &input); err != nil {
-			return jsonError("invalid input: " + err.Error()), true
-		}
-		cmd := sonarr.CommandRequest{
-			Name:     "SeriesSearch",
-			SeriesID: input.SeriesID,
-		}
-		if input.SeasonNumber != nil {
-			cmd.Name = "SeasonSearch"
-			cmd.SeasonNumber = input.SeasonNumber
-		}
-		result, err = a.sonarr.Command(ctx, cmd)
-
-	case "delete_series":
-		var input deleteSeriesInput
-		if err := json.Unmarshal(rawInput, &input); err != nil {
-			return jsonError("invalid input: " + err.Error()), true
-		}
-		err = a.sonarr.DeleteSeries(ctx, input.SeriesID, input.DeleteFiles)
-		if err == nil {
-			result = map[string]string{"status": "deleted"}
-		}
-
-	case "remove_blocklist_item":
-		var input removeBlocklistItemInput
-		if err := json.Unmarshal(rawInput, &input); err != nil {
-			return jsonError("invalid input: " + err.Error()), true
-		}
-		err = a.sonarr.DeleteBlocklistItem(ctx, input.ID)
-		if err == nil {
-			result = map[string]string{"status": "removed"}
-		}
-
-	case "grab_release":
-		var input grabReleaseInput
-		if err := json.Unmarshal(rawInput, &input); err != nil {
-			return jsonError("invalid input: " + err.Error()), true
-		}
-		result, err = a.sonarr.GrabRelease(ctx, input.GUID, input.IndexerID)
-
-	case "search_movies", "add_movie", "get_movie_queue", "get_movie_history", "check_movie_health", "remove_failed_movie",
-		"get_movie_detail", "get_movie_quality_profiles", "get_movie_root_folders", "get_movie_download_clients",
-		"get_movie_blocklist", "manual_movie_search", "update_movie_monitoring", "delete_movie",
-		"trigger_movie_search", "grab_movie_release", "remove_movie_blocklist_item":
-		if a.radarr == nil {
-			return jsonError("Radarr integration is not configured"), true
-		}
-		switch name {
-		case "search_movies":
-			var input searchMoviesInput
-			if err := json.Unmarshal(rawInput, &input); err != nil {
-				return jsonError("invalid input: " + err.Error()), true
-			}
-			result, err = a.radarr.Search(ctx, input.Term)
-		case "add_movie":
-			var input addMovieInput
-			if err := json.Unmarshal(rawInput, &input); err != nil {
-				return jsonError("invalid input: " + err.Error()), true
-			}
-			minAvail := input.MinimumAvailability
-			if minAvail == "" {
-				minAvail = "released"
-			}
-			result, err = a.radarr.Add(ctx, radarr.AddMovieRequest{
-				Title:               input.Title,
-				TMDBID:              input.TMDBID,
-				QualityProfileID:    input.QualityProfileID,
-				RootFolderPath:      input.RootFolderPath,
-				Monitored:           true,
-				MinimumAvailability: minAvail,
-			})
-		case "get_movie_queue":
-			result, err = a.radarr.Queue(ctx)
-		case "get_movie_history":
-			var input getMovieHistoryInput
-			if err := json.Unmarshal(rawInput, &input); err != nil {
-				return jsonError("invalid input: " + err.Error()), true
-			}
-			pageSize := input.PageSize
-			if pageSize == 0 {
-				pageSize = 20
-			}
-			result, err = a.radarr.History(ctx, pageSize)
-		case "check_movie_health":
-			result, err = a.radarr.Health(ctx)
-		case "remove_failed_movie":
-			var input removeFailedMovieInput
-			if err := json.Unmarshal(rawInput, &input); err != nil {
-				return jsonError("invalid input: " + err.Error()), true
-			}
-			err = a.radarr.RemoveFailed(ctx, input.ID, input.Blocklist)
-			if err == nil {
-				result = map[string]string{"status": "removed"}
-			}
-		case "get_movie_detail":
-			var input getMovieDetailInput
-			if err := json.Unmarshal(rawInput, &input); err != nil {
-				return jsonError("invalid input: " + err.Error()), true
-			}
-			result, err = a.radarr.GetMovie(ctx, input.MovieID)
-		case "get_movie_quality_profiles":
-			result, err = a.radarr.GetQualityProfiles(ctx)
-		case "get_movie_root_folders":
-			result, err = a.radarr.GetRootFolders(ctx)
-		case "get_movie_download_clients":
-			result, err = a.radarr.GetDownloadClients(ctx)
-		case "get_movie_blocklist":
-			var input getMovieBlocklistInput
-			if err := json.Unmarshal(rawInput, &input); err != nil {
-				return jsonError("invalid input: " + err.Error()), true
-			}
-			pageSize := input.PageSize
-			if pageSize == 0 {
-				pageSize = 20
-			}
-			result, err = a.radarr.GetBlocklist(ctx, pageSize)
-		case "manual_movie_search":
-			var input manualMovieSearchInput
-			if err := json.Unmarshal(rawInput, &input); err != nil {
-				return jsonError("invalid input: " + err.Error()), true
-			}
-			result, err = a.radarr.ManualSearch(ctx, input.MovieID)
-		case "update_movie_monitoring":
-			var input updateMovieMonitoringInput
-			if err := json.Unmarshal(rawInput, &input); err != nil {
-				return jsonError("invalid input: " + err.Error()), true
-			}
-			movie, err := a.radarr.GetMovie(ctx, input.MovieID)
-			if err != nil {
-				break
-			}
-			movie.Monitored = input.Monitored
-			result, err = a.radarr.UpdateMovie(ctx, movie)
-		case "delete_movie":
-			var input deleteMovieInput
-			if err := json.Unmarshal(rawInput, &input); err != nil {
-				return jsonError("invalid input: " + err.Error()), true
-			}
-			err = a.radarr.DeleteMovie(ctx, input.MovieID, input.DeleteFiles)
-			if err == nil {
-				result = map[string]string{"status": "deleted"}
-			}
-		case "trigger_movie_search":
-			var input triggerMovieSearchInput
-			if err := json.Unmarshal(rawInput, &input); err != nil {
-				return jsonError("invalid input: " + err.Error()), true
-			}
-			err = a.radarr.Command(ctx, radarr.CommandRequest{
-				Name:     radarr.CommandMoviesSearch,
-				MovieIDs: []int{input.MovieID},
-			})
-			if err == nil {
-				result = map[string]string{"status": "search triggered"}
-			}
-		case "grab_movie_release":
-			var input grabMovieReleaseInput
-			if err := json.Unmarshal(rawInput, &input); err != nil {
-				return jsonError("invalid input: " + err.Error()), true
-			}
-			err = a.radarr.GrabRelease(ctx, input.GUID, input.IndexerID)
-			if err == nil {
-				result = map[string]string{"status": "grabbed"}
-			}
-		case "remove_movie_blocklist_item":
-			var input removeMovieBlocklistItemInput
-			if err := json.Unmarshal(rawInput, &input); err != nil {
-				return jsonError("invalid input: " + err.Error()), true
-			}
-			err = a.radarr.DeleteBlocklistItem(ctx, input.ID)
-			if err == nil {
-				result = map[string]string{"status": "removed"}
-			}
-		}
-
-	case "list_indexers", "test_indexer", "test_all_indexers", "get_indexer_stats", "check_indexer_health", "search_indexers", "enable_indexer", "update_indexer_priority", "delete_indexer":
-		if a.prowlarr == nil {
-			return jsonError("Prowlarr integration is not configured"), true
-		}
-		switch name {
-		case "list_indexers":
-			result, err = a.prowlarr.ListIndexers(ctx)
-		case "test_indexer":
-			var input testIndexerInput
-			if err := json.Unmarshal(rawInput, &input); err != nil {
-				return jsonError("invalid input: " + err.Error()), true
-			}
-			err = a.prowlarr.TestIndexer(ctx, input.ID)
-			if err == nil {
-				result = map[string]string{"status": "ok"}
-			}
-		case "test_all_indexers":
-			result, err = a.prowlarr.TestAllIndexers(ctx)
-		case "get_indexer_stats":
-			result, err = a.prowlarr.GetIndexerStats(ctx)
-		case "check_indexer_health":
-			result, err = a.prowlarr.CheckHealth(ctx)
-		case "search_indexers":
-			var input searchIndexersInput
-			if err := json.Unmarshal(rawInput, &input); err != nil {
-				return jsonError("invalid input: " + err.Error()), true
-			}
-			result, err = a.prowlarr.Search(ctx, input.Query)
-		case "enable_indexer":
-			var input enableIndexerInput
-			if err := json.Unmarshal(rawInput, &input); err != nil {
-				return jsonError("invalid input: " + err.Error()), true
-			}
-			found, errStr := a.findIndexer(ctx, input.IndexerID)
-			if errStr != "" {
-				return errStr, true
-			}
-			found.Enable = input.Enabled
-			result, err = a.prowlarr.UpdateIndexer(ctx, found)
-		case "update_indexer_priority":
-			var input updateIndexerPriorityInput
-			if err := json.Unmarshal(rawInput, &input); err != nil {
-				return jsonError("invalid input: " + err.Error()), true
-			}
-			found, errStr := a.findIndexer(ctx, input.IndexerID)
-			if errStr != "" {
-				return errStr, true
-			}
-			found.Priority = input.Priority
-			result, err = a.prowlarr.UpdateIndexer(ctx, found)
-		case "delete_indexer":
-			var input deleteIndexerInput
-			if err := json.Unmarshal(rawInput, &input); err != nil {
-				return jsonError("invalid input: " + err.Error()), true
-			}
-			err = a.prowlarr.DeleteIndexer(ctx, input.IndexerID)
-			if err == nil {
-				result = map[string]string{"status": "deleted"}
-			}
-		}
-
-	case "list_requests", "approve_request", "decline_request", "get_request_detail", "delete_request", "retry_request", "get_request_count", "search_media":
-		if a.overseerr == nil {
-			return jsonError("Overseerr integration is not configured"), true
-		}
-		switch name {
-		case "list_requests":
-			var input listRequestsInput
-			if err := json.Unmarshal(rawInput, &input); err != nil {
-				return jsonError("invalid input: " + err.Error()), true
-			}
-			take := input.Take
-			if take == 0 {
-				take = 20
-			}
-			result, err = a.overseerr.ListRequests(ctx, input.Filter, take, input.Skip)
-		case "approve_request":
-			var input approveRequestInput
-			if err := json.Unmarshal(rawInput, &input); err != nil {
-				return jsonError("invalid input: " + err.Error()), true
-			}
-			err = a.overseerr.ApproveRequest(ctx, input.ID)
-			if err == nil {
-				result = map[string]string{"status": "approved"}
-			}
-		case "decline_request":
-			var input declineRequestInput
-			if err := json.Unmarshal(rawInput, &input); err != nil {
-				return jsonError("invalid input: " + err.Error()), true
-			}
-			err = a.overseerr.DeclineRequest(ctx, input.ID)
-			if err == nil {
-				result = map[string]string{"status": "declined"}
-			}
-		case "get_request_detail":
-			var input getRequestDetailInput
-			if err := json.Unmarshal(rawInput, &input); err != nil {
-				return jsonError("invalid input: " + err.Error()), true
-			}
-			result, err = a.overseerr.GetRequest(ctx, input.ID)
-		case "delete_request":
-			var input deleteRequestInput
-			if err := json.Unmarshal(rawInput, &input); err != nil {
-				return jsonError("invalid input: " + err.Error()), true
-			}
-			err = a.overseerr.DeleteRequest(ctx, input.ID)
-			if err == nil {
-				result = map[string]string{"status": "deleted"}
-			}
-		case "retry_request":
-			var input retryRequestInput
-			if err := json.Unmarshal(rawInput, &input); err != nil {
-				return jsonError("invalid input: " + err.Error()), true
-			}
-			result, err = a.overseerr.RetryRequest(ctx, input.ID)
-		case "get_request_count":
-			result, err = a.overseerr.GetRequestCount(ctx)
-		case "search_media":
-			var input searchMediaInput
-			if err := json.Unmarshal(rawInput, &input); err != nil {
-				return jsonError("invalid input: " + err.Error()), true
-			}
-			page := input.Page
-			if page == 0 {
-				page = 1
-			}
-			result, err = a.overseerr.SearchMedia(ctx, input.Query, page)
-		}
-
-	case "notebook_write", "notebook_read", "notebook_list", "notebook_delete":
-		if a.notebook == nil {
-			return jsonError("Notebook integration is not configured"), true
-		}
-		switch name {
-		case "notebook_write":
-			var input notebookWriteInput
-			if err := json.Unmarshal(rawInput, &input); err != nil {
-				return jsonError("invalid input: " + err.Error()), true
-			}
-			noteType := notebook.NoteType(input.Type)
-			if noteType != notebook.Pinned && noteType != notebook.Reference {
-				return jsonError("type must be 'pinned' or 'reference'"), true
-			}
-			// Check for duplicate titles to avoid creating redundant notes.
-			if input.ID == "" {
-				existing, searchErr := a.notebook.Search(ctx, input.Title)
-				if searchErr == nil {
-					for _, n := range existing {
-						if strings.EqualFold(n.Title, input.Title) {
-							data, _ := json.Marshal(map[string]string{
-								"warning":     "a note with a similar title already exists",
-								"existing_id": n.ID,
-								"title":       n.Title,
-							})
-							return string(data), false
-						}
-					}
-				}
-			}
-			err = a.notebook.Write(ctx, notebook.Note{
-				ID:      input.ID,
-				Type:    noteType,
-				Title:   input.Title,
-				Content: input.Content,
-			})
-			if err == nil {
-				result = map[string]string{"status": "saved"}
-			}
-		case "notebook_read":
-			var input notebookReadInput
-			if err := json.Unmarshal(rawInput, &input); err != nil {
-				return jsonError("invalid input: " + err.Error()), true
-			}
-			result, err = a.notebook.Read(ctx, input.ID)
-		case "notebook_list":
-			var input notebookListInput
-			if err := json.Unmarshal(rawInput, &input); err != nil {
-				return jsonError("invalid input: " + err.Error()), true
-			}
-			summaries, listErr := a.notebook.List(ctx)
-			if listErr != nil {
-				err = listErr
-			} else if input.Type != "" {
-				filtered := make([]notebook.NoteSummary, 0)
-				for _, s := range summaries {
-					if string(s.Type) == input.Type {
-						filtered = append(filtered, s)
-					}
-				}
-				result = filtered
-			} else {
-				result = summaries
-			}
-		case "notebook_delete":
-			var input notebookDeleteInput
-			if err := json.Unmarshal(rawInput, &input); err != nil {
-				return jsonError("invalid input: " + err.Error()), true
-			}
-			err = a.notebook.Delete(ctx, input.ID)
-			if err == nil {
-				result = map[string]string{"status": "deleted"}
-			}
-		}
-
-	default:
-		return jsonError("unknown tool: " + name), true
+	if result, isErr, handled := a.dispatchSonarr(ctx, name, rawInput); handled {
+		return result, isErr
 	}
-
-	if err != nil {
-		a.logger.Warn("tool error", "tool", name, "error", err)
-		return jsonError(err.Error()), true
+	if result, isErr, handled := a.dispatchRadarr(ctx, name, rawInput); handled {
+		return result, isErr
 	}
-
-	data, marshalErr := json.Marshal(result)
-	if marshalErr != nil {
-		return jsonError("failed to marshal result: " + marshalErr.Error()), true
+	if result, isErr, handled := a.dispatchProwlarr(ctx, name, rawInput); handled {
+		return result, isErr
 	}
-	return string(data), false
+	if result, isErr, handled := a.dispatchOverseerr(ctx, name, rawInput); handled {
+		return result, isErr
+	}
+	if result, isErr, handled := a.dispatchNotebook(ctx, name, rawInput); handled {
+		return result, isErr
+	}
+	return jsonError("unknown tool: " + name), true
 }
 
 // findIndexer fetches the indexer list from Prowlarr and returns the indexer with the given ID.
